% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_team.R
\name{geom_big_games}
\alias{geom_big_games}
\title{geom_big_games}
\usage{
geom_big_games(df, team_name, team_color)
}
\arguments{
\item{df}{A dataframe resulting from calling the marble_game funciton}

\item{team_name}{NCAA D1 FBS Football team name}

\item{team_color}{Color to represent team}
}
\value{
A ggplot geom plotting the most consequential marble changes for a team in a  given season.
}
\description{
geom_big_games
}
