% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_team.R
\name{geom_big_loss}
\alias{geom_big_loss}
\title{geom_big_loss}
\usage{
geom_big_loss(df, team_name, team_color)
}
\arguments{
\item{df}{A dataframe resulting from calling the marble_game funciton}

\item{team_name}{NCAA D1 FBS Football team name}

\item{team_color}{Color to represent team}
}
\value{
A ggplot geom plotting the biggest loss for a team or any loss that results in losing 40+ marbles.
}
\description{
geom_big_loss
}
