% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_team.R
\name{geom_big_wins}
\alias{geom_big_wins}
\title{geom_big_wins}
\usage{
geom_big_wins(df, team_name, team_color)
}
\arguments{
\item{df}{A dataframe resulting from calling the marble_game funciton}

\item{team_name}{NCAA D1 FBS Football team name}

\item{team_color}{Color to represent team}
}
\value{
A ggplot geom plotting a team's largest win and any win that results in a gain of 40+ marbles.
}
\description{
geom_big_wins
}
