% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_team.R
\name{geom_team}
\alias{geom_team}
\title{geom_team}
\usage{
geom_team(df, team_name, team_color)
}
\arguments{
\item{df}{A dataframe resulting from calling the marble_game funciton}

\item{team_name}{NCAA D1 FBS Football team name}

\item{team_color}{Color to represent team}
}
\value{
A ggplot geom plotting the marbles of a team for a given season.
}
\description{
geom_team
}
