% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{get_season}
\alias{get_season}
\title{get_season}
\usage{
get_season(df, yr = 2017, wk = 36)
}
\arguments{
\item{df}{Data frame of game data - should always be `ncaa_games`}

\item{yr}{Number indicating year of NCAA football season}

\item{wk}{Number indicating maximum week within NCAA football season}
}
\value{
A dataframe with NCAA FBS football game results
}
\description{
get_season
}
