% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marbl.R
\name{marble_game}
\alias{marble_game}
\title{marble_game}
\usage{
marble_game(df, yr = 2017, wk = 36, p5_value = 120)
}
\arguments{
\item{df}{Data frame of NCAA football game results, should always be ncaa_games}

\item{yr}{Number indicating year of NCAA football season. Defaults to 2017.}

\item{wk}{Number indicating maximum week of NCAA football season. Defaults to 36 to ensure all weeks are included in a standard marble game.}

\item{p5_value}{Number indicating the amount of marbles for Power 5 conference teams have to start a season. Defaults to 120, while all other teams start with 100.}
}
\value{
A dataframe of game data.
}
\description{
marble_game
}
