% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p5.R
\docType{data}
\name{p5}
\alias{p5}
\title{NCAA Power 5 Conference Teams, 2017}
\format{A dataframe with 65 rows and 2 variables:
\describe{
  \item{team}{Team name}
  \item{conf}{Conference name}
}}
\source{
\url{https://en.wikipedia.org/wiki/Power_Five_conferences}
}
\usage{
p5
}
\description{
NCAA Power 5 Conference Teams, 2017
}
\keyword{datasets}
