% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marble_plot.R
\name{marble_plot}
\alias{marble_plot}
\title{marble_plot}
\usage{
marble_plot(df, alpha_level = 0.4)
}
\arguments{
\item{df}{A dataframe resulting from calling the marble_game funciton}
}
\value{
A ggplot object plotting the results of a marble_game() call
}
\description{
marble_plot
}
